new player search bar component
