+++ app/public/src/pages/component/moderation/moderation-panel.css
@@ -1,5 +1,6 @@
 .moderation-panel {
   width: min(80vw, 900px);
+  min-height: 60vh;
   color: var(--color-fg-primary);
 }
