+++ app/public/src/pages/component/profile/profile.css
@@ -229,31 +229,6 @@ ul.titles > li.selected .title-name {
   color: var(--color-fg-gold);
 }
 
-.search-suggestions {
-  display: flex;
-  padding: 0;
-  gap: 10px;
-  flex-wrap: wrap;
-  justify-content: space-around;
-  list-style: none;
-}
-
-.search-suggestions .player.my-box {
-  display: flex;
-  flex-flow: column;
-  padding: 5px;
-}
-
-.search-suggestions .player.my-box.banned {
-  color: red;
-}
-
-.search-suggestions .player.my-box img {
-  width: 40px;
-  height: 40px;
-  margin: 0 auto;
-}
-
 .player-history .actions {
   font-size: 80%;
   display: flex;
