+++ app/public/src/pages/component/profile/profile.tsx
@@ -1,13 +1,11 @@
-import firebase from "firebase/compat/app"
 import type React from "react"
-import { useCallback, useRef, useState } from "react"
+import { useState } from "react"
 import { useTranslation } from "react-i18next"
 import { Tab, TabList, TabPanel, Tabs } from "react-tabs"
 import type { IGameRecord } from "../../../../../models/colyseus-models/game-record"
-import { type ISuggestionUser, Role, Title } from "../../../../../types"
-import { debounce } from "../../../../../utils/function"
+import { Role, Title } from "../../../../../types"
 import { keys } from "../../../../../utils/object"
-import { useAppDispatch, useAppSelector } from "../../../hooks"
+import { useAppSelector } from "../../../hooks"
 import {
   ban,
   giveBooster,
@@ -16,7 +14,7 @@ import {
   searchById,
   unban
 } from "../../../network"
-import { setSearchedUser } from "../../../stores/LobbyStore"
+import { PlayerSearchBar } from "../search/player-search-bar"
 import { AccountTab } from "./account-tab"
 import { AvatarTab } from "./avatar-tab"
 import { EloTab } from "./elo-tab"
@@ -25,79 +23,17 @@ import GameHistory from "./game-history"
 import PlayerBox from "./player-box"
 import { ProfileChatHistory } from "./profile-chat-history"
 import { ProgressTab } from "./progress-tab"
-import { SearchBar } from "./search-bar"
-import SearchResults from "./search-results"
 import { TitleTab } from "./title-tab"
 import "./profile.css"
 
 export default function Profile() {
   const { t } = useTranslation()
-  const dispatch = useAppDispatch()
   const user = useAppSelector((state) => state.network.profile)
-  const [suggestions, setSuggestions] = useState<ISuggestionUser[]>([])
   const searchedUser = useAppSelector((state) => state.lobby.searchedUser)
 
   const profile = searchedUser ?? user
   const [gameHistory, setGameHistory] = useState<IGameRecord[]>([])
   const [rightPanel, setRightPanel] = useState<"chat" | "game">("game")
-
-  const [loading, setLoading] = useState<boolean>(false)
-  const [error, setError] = useState<string>("")
-  const abortControllerRef = useRef<AbortController | null>(null)
-
-  async function searchName(query: string) {
-    abortControllerRef.current = new AbortController()
-    const { signal } = abortControllerRef.current
-    setLoading(true)
-    setError("")
-    try {
-      const token = await firebase.auth().currentUser?.getIdToken()
-      const res = await fetch(`/players?name=${encodeURIComponent(query)}`, {
-        method: "GET",
-        headers: {
-          "Content-Type": "application/json",
-          Authorization: `Bearer ${token}`
-        },
-        signal
-      })
-      if (res.ok) {
-        const suggestions = await res.json()
-        if (suggestions.length === 0) {
-          setError(t("no_results_found"))
-        } else {
-          setSuggestions(suggestions)
-          setError("")
-        }
-      } else {
-        setError(res.statusText)
-      }
-    } catch (err: any) {
-      if (err.name !== "AbortError") {
-        setError(err.message)
-      }
-    }
-    setLoading(false)
-  }
-
-  const debouncedSearchName = useRef(debounce(searchName, 500)).current
-
-  function onSearchQueryChange(query: string) {
-    abortControllerRef.current?.abort()
-    if (query) {
-      debouncedSearchName(query)
-    } else {
-      resetSearch()
-    }
-  }
-
-  const resetSearch = useCallback(
-    (user = searchedUser) => {
-      dispatch(setSearchedUser(user))
-      setSuggestions([])
-      setError("")
-    },
-    [dispatch]
-  )
 
   return (
     <div className="profile-modal">
@@ -108,22 +44,14 @@ export default function Profile() {
         {profile && <PlayerBox user={profile} history={gameHistory} />}
       </div>
 
-      <SearchBar onChange={onSearchQueryChange} />
+      <PlayerSearchBar
+        onSelect={(suggestion) => {
+          searchById(suggestion.id)
+        }}
+      />
 
       <div className="profile-actions">
-        {loading ? (
-          <div className="loading">{t("loading")}</div>
-        ) : error ? (
-          <div className="error">{error}</div>
-        ) : suggestions.length > 0 ? (
-          <SearchResults
-            suggestions={suggestions}
-            onSelect={(suggestion) => {
-              resetSearch()
-              searchById(suggestion.id)
-            }}
-          />
-        ) : searchedUser ? (
+        {searchedUser ? (
           <OtherProfileActions
             rightPanel={rightPanel}
             setRightPanel={setRightPanel}
